requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0
opencv-python>=4.7.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
openai>=1.0.0

# API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Logging (optional YAML support)
pyyaml>=6.0  # Optional, for YAML config files

# Security & Rate Limiting
slowapi>=0.1.9  # Rate limiting for FastAPI
pydantic-settings>=2.0.0  # Settings management

# Optional dependencies
# efficientnet-pytorch>=0.7.1  # For better character feature extraction

